+++ CustomerVehicleManagement.Api/Persons/IPersonRepository.cs
@@ -10,12 +10,12 @@ namespace CustomerVehicleManagement.Api.Persons
         Task AddPersonAsync(Person entity);
         void DeletePerson(Person entity);
         void FixTrackingState();
-        Task<bool> PersonExistsAsync(int id);
+        Task<bool> PersonExistsAsync(long id);
         void UpdatePersonAsync(Person personUpdateDto);
         Task<IReadOnlyList<PersonReadDto>> GetPersonsAsync();
         Task<IReadOnlyList<PersonInListDto>> GetPersonsListAsync();
-        Task<PersonReadDto> GetPersonAsync(int id);
-        Task<Person> GetPersonEntityAsync(int id);
+        Task<PersonReadDto> GetPersonAsync(long id);
+        Task<Person> GetPersonEntityAsync(long id);
         Task<bool> SaveChangesAsync();
     }
 }
