+++ Menominee.UiExperiments/Services/PersonDataService.cs
@@ -45,7 +45,7 @@ namespace Menominee.UiExperiments.Services
             return null;
         }
 
-        public async Task<PersonReadDto> GetPersonDetails(int id)
+        public async Task<PersonReadDto> GetPersonDetails(long id)
         {
             try
             {
@@ -74,7 +74,7 @@ namespace Menominee.UiExperiments.Services
             //}
         }
 
-        public async Task DeletePerson(int id)
+        public async Task DeletePerson(long id)
         {
             try
             {
